Clean up architecture: Remove old and comparison references
- Replace agents_v2.py with agents.py as the standard implementation
- Remove mcp_bridge.py (no longer needed)
- Rename test_simplified_agents_tdd.py to test_agents.py
- Remove test_simplified_vs_original.py comparison test
- Rename demo_simplified_agents.py to demo_agents.py
- Rename SIMPLIFIED_ARCHITECTURE_GUIDE.md to ARCHITECTURE_GUIDE.md
- Remove MIGRATION_NOTES.md
- Update all imports from agents_v2 to agents
- Update test files to use new factory pattern with dependency injection
- Update CLAUDE.md to remove simplified architecture references
- Update documentation to remove comparison language and migration guides

The architecture is now the standard without legacy references:
- Clean factory functions with dependency injection
- Type-safe Pydantic response models
- Centralized error handling via pydantic-ai
- Direct MCP server access
- Follows pydantic-ai best practices
