+++ CLAUDE.md
@@ -32,7 +32,7 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 - **Agent integration tests**: `uv run pytest tests/agents/test_ai_agent_integration.py -v`
 - **Python execution agent tests**: `uv run pytest tests/agents/test_python_exec_agent_integration.py -v`
 - **Script generation agent tests**: `uv run pytest tests/agents/test_script_agent_integration.py -v`
-- **Simplified architecture TDD tests**: `uv run pytest tests/test_simplified_agents_tdd.py -v`
+- **Agent architecture tests**: `uv run pytest tests/test_agents.py -v`
 
 #### Specific Test Classes and Methods
 - **Run specific test class**: `uv run pytest tests/tools/test_loadset_enhanced.py::TestLoadSetReadJson -v`
@@ -43,16 +43,16 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 - **MCP tests (80)**: Mock MCP server tests for tool functionality in `tests/mcps/`
 - **Agent tests (30)**: Expensive LLM/AI agent tests requiring API keys in `tests/agents/` - run explicitly with `-m expensive`
 - **Visual tests (1)**: Chart generation tests in `tests/agents/` - run explicitly with `-m visuals`
-- **Simplified architecture tests (19)**: TDD tests for new simplified agent architecture
-- **Total fast tests (156)**: Core + MCP + TDD tests that run by default
+- **Agent architecture tests (19)**: TDD tests for pydantic-ai agent architecture
+- **Total fast tests (156)**: Core + MCP + Agent tests that run by default
 - **Total tests (187)**: All tests including expensive and visual tests
 
 ### Verification
 - **Verify setup**: `uv run python verify_setup.py`
 
-## Simplified Agent Architecture (Recommended)
+## Agent Architecture
 
-This project now uses a **simplified pydantic-ai architecture** that achieves a **57.2% code reduction** while following best practices and maintaining full functionality.
+This project uses a **pydantic-ai architecture** following best practices with dependency injection and type-safe responses.
 
 ### Model Selection
 Set a single environment variable to choose your AI model:
@@ -70,9 +70,9 @@ export AI_MODEL="openai:gpt-4o"
 export AI_MODEL="google-gla:gemini-1.5-flash"
 ```
 
-### Simple Agent Usage
+### Agent Usage
 ```python
-from tools.agents_v2 import create_loadset_agent, create_python_agent, create_script_agent
+from tools.agents import create_loadset_agent, create_python_agent, create_script_agent
 from tools.dependencies import MCPServerProvider
 
 # Create agent and dependencies once
@@ -100,13 +100,12 @@ result = await script_agent.run(
 )
 ```
 
-### Testing the Simplified Architecture
-- **Test TDD implementation**: `uv run pytest tests/test_simplified_agents_tdd.py -v`
-- **Compare architectures**: `uv run python test_simplified_vs_original.py`
+### Testing the Agent Architecture
+- **Test agent implementation**: `uv run pytest tests/test_agents.py -v`
 - **All fast tests**: `uv run pytest tests/tools/ tests/mcps/ -v`
 
 ### Benefits
-- **57.2% code reduction**: From 400+ lines to 171 lines
+- **Clean architecture**: Factory functions with dependency injection
 - **Type-safe responses**: Pydantic models instead of raw dictionaries
 - **True dependency injection**: Using RunContext for MCP server access
 - **Centralized error handling**: No manual try-catch in tools
@@ -114,7 +113,7 @@ result = await script_agent.run(
 - **Zero abstraction overhead**: Direct MCP server access
 
 ### Architecture Documentation
-See `SIMPLIFIED_ARCHITECTURE_GUIDE.md` for comprehensive documentation, migration guide, and best practices.
+See `ARCHITECTURE_GUIDE.md` for comprehensive documentation and best practices.
 
 ## FIREWORKS AI Integration (Legacy)
