--- MIGRATION_NOTES.md
@@ -1,97 +1,0 @@
-# Clean Agent Architecture Migration
-
-## ✅ Migration Complete
-
-The project has been successfully migrated from boilerplate agent classes to a clean pydantic-ai architecture.
-
-## 🆕 New Clean Architecture Files
-
-### Core Files (Use These)
-- `tools/model_config.py` - Single environment-based model selection
-- `tools/mcp_bridge.py` - Clean MCP server integration layer  
-- `tools/agents.py` - Global agents with tool decorators (90% less code)
-
-### Demo and Test Files
-- `demo_clean_agents.py` - Demonstration of new clean API
-- `test_simple_agent.py` - Basic functionality test
-- `test_clean_agent_integration.py` - Comprehensive integration tests
-
-## 📦 Legacy Files (Can Be Removed)
-
-These files contain the old boilerplate patterns and are no longer needed:
-
-### Old Agent Classes (650+ lines → 0 lines needed)
-- `tools/fireworks_mcp_agent.py` - FireworksMCPAgent class (300+ lines)
-- `tools/loads_agent.py` - Old LoadSet agent implementation  
-- `tools/script_agent_client.py` - Old script agent client
-- `demo_fireworks_agent.py` - Old FIREWORKS demo with boilerplate
-- `test_fireworks_mcp_integration.py` - Old integration tests
-- `test_fireworks_integration.py` - Old basic tests
-
-### Keep But Mark as Legacy
-- `tools/fireworks_client.py` - Still useful for direct model access
-- `tests/agents/test_ai_agent_integration.py` - Can be updated to use new agents
-
-## 🔄 Migration Benefits Achieved
-
-### Before (Old Architecture)
-```python
-# Required custom agent classes
-from fireworks_mcp_agent import FireworksMCPAgent
-from mcps.loads_mcp_server import create_mcp_server
-
-server = create_mcp_server() 
-agent = FireworksMCPAgent(server, model_name="...")
-result = await agent.process_user_prompt("Load and process data")
-```
-
-### After (Clean Architecture)  
-```python
-# Zero boilerplate - direct agent usage
-from tools.agents import loadset_agent
-
-result = await loadset_agent.run("Load and process data")
-```
-
-## 📊 Code Reduction Summary
-
-- **FireworksMCPAgent**: 300+ lines → 0 lines (eliminated)
-- **FireworksPythonExecutionAgent**: 150+ lines → 0 lines (eliminated)
-- **AnthropicMCPTestAgent**: 200+ lines → 0 lines (eliminated)
-- **Total boilerplate eliminated**: 650+ lines
-- **New core implementation**: ~100 lines total
-- **Code reduction**: 90%+
-
-## 🎯 Usage Patterns
-
-### Model Selection (Single Environment Variable)
-```bash
-export AI_MODEL="anthropic:claude-3-5-sonnet-latest"  # Anthropic
-export AI_MODEL="fireworks:accounts/fireworks/models/llama-v3p3-70b-instruct"  # FIREWORKS
-export AI_MODEL="openai:gpt-4o"  # OpenAI
-```
-
-### Agent Usage (Zero Boilerplate)
-```python
-from tools.agents import loadset_agent, python_agent, script_agent
-
-# LoadSet processing
-result = await loadset_agent.run("Load solution/loads/new_loads.json and convert to kN")
-
-# Python execution  
-result = await python_agent.run("Execute: print('Hello World')")
-
-# Script generation
-result = await script_agent.run("Generate a data analysis script")
-```
-
-## ✅ All Benefits Delivered
-
-1. **90% code reduction** - From 650+ lines to ~100 lines
-2. **Single model selection** - Just set AI_MODEL environment variable
-3. **Provider independence** - MCP completely separate from LLM choice
-4. **Pydantic-AI best practices** - Global agents with tool decorators
-5. **Zero boilerplate** - Direct `.run()` calls, no custom classes
-6. **Easy extensibility** - Add new tools with simple decorators
-
-The clean architecture is production-ready and significantly easier to maintain and extend.
