--- test_simplified_vs_original.py
@@ -1,114 +1,0 @@
-"""
-Comparison test between original and simplified agent architectures.
-
-This test demonstrates that the simplified architecture achieves the same
-functionality as the original while using significantly less code.
-"""
-
-import asyncio
-import os
-import sys
-from pathlib import Path
-from dotenv import load_dotenv
-
-# Add project root to path
-project_root = Path(__file__).parent
-sys.path.insert(0, str(project_root))
-
-# Load environment variables
-load_dotenv()
-
-# Import original architecture
-from tools.agents import loadset_agent as original_loadset_agent
-from tools.model_config import validate_model_config
-
-# Import simplified architecture
-from tools.agents_v2 import create_loadset_agent
-from tools.dependencies import MCPServerProvider
-
-
-async def test_original_architecture():
-    """Test the original agent architecture."""
-    print("🔍 Testing Original Architecture...")
-    
-    try:
-        result = await original_loadset_agent.run(
-            "Load 'solution/loads/new_loads.json' and give me a summary of the LoadSet"
-        )
-        print(f"✅ Original agent worked: {len(result.output)} character response")
-        return True
-    except Exception as e:
-        print(f"❌ Original agent failed: {e}")
-        return False
-
-
-async def test_simplified_architecture():
-    """Test the simplified agent architecture."""
-    print("🔍 Testing Simplified Architecture...")
-    
-    try:
-        # Create agent and dependencies
-        agent = create_loadset_agent()
-        deps = MCPServerProvider()
-        
-        result = await agent.run(
-            "Load 'solution/loads/new_loads.json' and give me a summary of the LoadSet",
-            deps=deps
-        )
-        print(f"✅ Simplified agent worked: {len(result.output)} character response")
-        return True
-    except Exception as e:
-        print(f"❌ Simplified agent failed: {e}")
-        return False
-
-
-async def compare_architectures():
-    """Compare both architectures side by side."""
-    print("🏗️ Architecture Comparison Test")
-    print("=" * 50)
-    
-    # Validate configuration
-    is_valid, error = validate_model_config()
-    if not is_valid:
-        print(f"❌ Configuration error: {error}")
-        return
-    
-    # Test both architectures
-    original_success = await test_original_architecture()
-    simplified_success = await test_simplified_architecture()
-    
-    print("\n📊 Results:")
-    print(f"Original Architecture: {'✅ PASS' if original_success else '❌ FAIL'}")
-    print(f"Simplified Architecture: {'✅ PASS' if simplified_success else '❌ FAIL'}")
-    
-    if original_success and simplified_success:
-        print("\n🎉 Both architectures work correctly!")
-        
-        # Show code metrics
-        print("\n📈 Code Metrics:")
-        
-        # Original architecture lines (approximate)
-        original_lines = 400  # From agents.py + mcp_bridge.py
-        
-        # Simplified architecture lines
-        simplified_lines = 171  # From agents_v2.py
-        
-        reduction = ((original_lines - simplified_lines) / original_lines) * 100
-        
-        print(f"Original Architecture: ~{original_lines} lines")
-        print(f"Simplified Architecture: {simplified_lines} lines")
-        print(f"Code Reduction: {reduction:.1f}%")
-        
-        print("\n🏆 Simplified architecture achievements:")
-        print("• Eliminated MCP bridge abstraction")
-        print("• Implemented dependency injection")
-        print("• Added structured Pydantic response models")
-        print("• Centralized error handling")
-        print("• Reduced boilerplate by 57%")
-        
-    else:
-        print("\n⚠️ One or both architectures failed")
-
-
-if __name__ == "__main__":
-    asyncio.run(compare_architectures())
